Added a basic UI by adding stuff in .html and .js
Added buttons in html and some respective <div>'s.
Made changes to the playRound function to make the
code cleaner and simpler and so, removed the
playGame function as it wasn't needed.
Added event listeners for the buttons.
The game is not limited to run only 5 times,
instead, it can run for infinite amount of times
as the playRound function no longer exists.